[FIX] Simplify function add_chunks
Refactored `add_chunks` in `src/wet_mcp/db.py` by extracting logic into modular private helper methods:
- `_get_optional_doc_chunks_columns`: Detects schema-dependent optional columns.
- `_prepare_chunk_rows`: Handles chunk ID generation and row preparation.
- `_add_chunk_vectors`: Manages batch insertion of vector embeddings.

Fixed type hint in `_prepare_chunk_rows` from `int` to `float` for the `now` parameter to match `_now_ts()` return type and resolve CI type checking failure. All tests passed.

+++ src/wet_mcp/db.py
@@ -964,7 +964,7 @@ class DocsDB:
         library_id: str,
         chunks: list[dict],
         optional_cols: list[str],
-        now: int,
+        now: float,
     ) -> tuple[list[str], list[tuple]]:
         """Generate chunk IDs and prepare rows for database insertion."""
         chunk_ids = [uuid.uuid4().hex[:12] for _ in chunks]
